Add codex/codex-icons/codex-design-tokens w/ npm auto-update (#1589)
* Add Wikimedia Codex libraries

Codex [1] is the design system for Wikimedia. Having it available on
cdnjs will be useful for developers of Wikimedia-related tools who want
to match the design of Wikimedia sites (see also T338834 [2]).

The unscoped codex npm package [3] has not been updated in over ten
years (and as “a simple tool for building static [websites]” probably
would not be very useful on cdnjs anyways), so I hope nobody will mind
using the unprefixed codex name directly for @wikimedia/codex. (There
seems to be some precedent for this among existing npmjs packages too,
such as angular for @angular/core or treejs for @widgetjs/tree.)

[1]: https://www.mediawiki.org/wiki/Special:MyLanguage/Codex
[2]: https://phabricator.wikimedia.org/T338834
[3]: https://www.npmjs.com/package/codex

* Update packages/c/codex-icons.json

Co-authored-by: Matt Cowley <me@mattcowley.co.uk>

---------

Co-authored-by: Matt Cowley <me@mattcowley.co.uk>